+++ internal/test_android_21.sh
@@ -8,11 +8,6 @@ cd $rootDir
 
 # The following modules support running tests on Android 21
 ./gradlew \
-  zoomimage-compose-coil2:connectedAndroidTest \
-  zoomimage-compose-coil2-core:connectedAndroidTest \
-  zoomimage-compose-glide:connectedAndroidTest \
-  zoomimage-compose-sketch3:connectedAndroidTest \
-  zoomimage-compose-sketch3-core:connectedAndroidTest \
   zoomimage-core:connectedAndroidTest \
   zoomimage-core-coil2:connectedAndroidTest \
   zoomimage-core-coil3:connectedAndroidTest \